# Improve performance, fix wall crawler tile weight table bug