Wall crawler had one row that was incorrectly duplicated: 
`1, 0.9, 0.75, 0.6, 0.5, 0.4, 0.3, 0.25, 0.25, 0.25, 0.25, 0.25, 0.25, 0.3, 0.4, 0.5, 0.6, 0.75, 0.9, 1,`
whereas the middle row should have been duplicated instead: 
`1, 0.9, 0.75, 0.6, 0.5, 0.4, 0.3, 0.25, 0.1, 0, 0, 0.1, 0.25, 0.3, 0.4, 0.5, 0.6, 0.75, 0.9, 1,`

Afterwards, I converted the array to integers get rid of floating point inconsistencies for testing the next step: 

`TileWeightEngine` does not need to play moves to evaluate them. The only thing it cares about is how much the sum of move's tiles add up to in its weight table, and the rest of the game state doesn't matter. By looping through the move's tiles instead of making the move, evaluating the entire board, and unmaking the move, we get a speed-up of around 200x on my machine. 

Previous approach: 
```
Game  1: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  2: WallCrawler    Turtle         Turtle         WallCrawler      Scores: [74, 70, 64, 77]  Winner(s): WallCrawler
Game  3: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  4: Turtle         WallCrawler    Turtle         WallCrawler      Scores: [75, 77, 79, 73]  Winner(s): Turtle
Game  5: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  6: Turtle         Turtle         WallCrawler    WallCrawler      Scores: [76, 77, 74, 77]  Winner(s): Turtle, WallCrawler
Game  7: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  8: Turtle         Turtle         WallCrawler    WallCrawler      Scores: [76, 77, 74, 77]  Winner(s): Turtle, WallCrawler
Game  9: Turtle         WallCrawler    WallCrawler    Turtle           Scores: [73, 76, 73, 68]  Winner(s): WallCrawler
Game 10: WallCrawler    Turtle         Turtle         WallCrawler      Scores: [74, 70, 64, 77]  Winner(s): WallCrawler

Ranking by elo_end desc:
Rank Name                       Elo  Games      Score  Avg Score   Wins  Win Rate
   0 WallCrawler                 15     10        751      75.10      5    50.00%
   1 WallCrawler                  5     10        740      74.00      4    40.00%
   2 Turtle                      -3     10        740      74.00      4    40.00%
   3 Turtle                     -17     10        737      73.70      3    30.00%

Tournament ran for 201.5269s with avg match duration 20.1505s
```

New approach: 
```
Game  1: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  2: WallCrawler    Turtle         Turtle         WallCrawler      Scores: [74, 70, 64, 77]  Winner(s): WallCrawler
Game  3: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  4: Turtle         WallCrawler    Turtle         WallCrawler      Scores: [75, 77, 79, 73]  Winner(s): Turtle
Game  5: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  6: Turtle         Turtle         WallCrawler    WallCrawler      Scores: [76, 77, 74, 77]  Winner(s): Turtle, WallCrawler
Game  7: WallCrawler    WallCrawler    Turtle         Turtle           Scores: [78, 69, 78, 74]  Winner(s): WallCrawler, Turtle
Game  8: Turtle         Turtle         WallCrawler    WallCrawler      Scores: [76, 77, 74, 77]  Winner(s): Turtle, WallCrawler
Game  9: Turtle         WallCrawler    WallCrawler    Turtle           Scores: [73, 76, 73, 68]  Winner(s): WallCrawler
Game 10: WallCrawler    Turtle         Turtle         WallCrawler      Scores: [74, 70, 64, 77]  Winner(s): WallCrawler

Ranking by elo_end desc:
Rank Name                       Elo  Games      Score  Avg Score   Wins  Win Rate
   0 WallCrawler                 15     10        751      75.10      5    50.00%
   1 WallCrawler                  5     10        740      74.00      4    40.00%
   2 Turtle                      -3     10        740      74.00      4    40.00%
   3 Turtle                     -17     10        737      73.70      3    30.00%

Tournament ran for 0.9931s with avg match duration 0.0951s
```